pyproject.toml
[project]
name = "nano-banana-lab"
version = "2.0.0"
description = "AI Image Generation API powered by Google Gemini"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "HyxiaoGe" }
]
keywords = ["ai", "image-generation", "gemini", "fastapi", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "passlib[bcrypt]>=1.7.4",
    # Cache & Queue (Required)
    "redis[hiredis]>=5.0.0",
    # Task Queue (Required)
    "arq>=0.26.0",
    # WebSocket
    "websockets>=12.0",
    # AI & Image Processing
    "google-genai>=1.0.0",
    "Pillow>=10.0.0",
    # Cloud Storage
    "boto3>=1.34.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "aiofiles>=25.1.0",
    "minio>=7.2.20",
    "sqlalchemy>=2.0.46",
    "asyncpg>=0.31.0",
    "alembic>=1.18.3",
    "greenlet>=3.3.1",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    # Linting & Formatting
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    # Type stubs
    "types-redis>=4.6.0",
    "types-aiofiles>=24.1.0",
    # Security
    "bandit[toml]>=1.7.10",
]

[project.urls]
Homepage = "https://github.com/HyxiaoGe/nano-banana-lab"
Repository = "https://github.com/HyxiaoGe/nano-banana-lab"
Documentation = "https://github.com/HyxiaoGe/nano-banana-lab#readme"

[project.scripts]
nano-banana = "api.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "services", "core", "database", "i18n"]

# ============ Tool Configurations ============

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = ["scripts/"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "C901",   # too complex
    "ARG001", # unused function argument (common in FastAPI deps)
    "ARG002", # unused method argument (common in FastAPI deps)
    "SIM117", # nested with statements (clearer in tests)
    "SIM102", # nested if statements (sometimes clearer)
    "SIM105", # contextlib.suppress (explicit try/except clearer)
    "E741",   # ambiguous variable name (l, O, I)
]

[tool.ruff.lint.isort]
known-first-party = ["api", "services", "core", "database", "i18n"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_ignores = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
ignore_missing_imports = true
no_implicit_optional = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=api --cov=services --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["api", "services", "core", "database"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

# ============ Security ============

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv", "node_modules"]
skips = [
    "B101",  # assert_used (ok in tests)
    "B110",  # try_except_pass (intentional in some async code)
    "B324",  # hashlib md5 (used for cache keys, not security)
]
